+++ src/main/kotlin/ru/hse/sd/hwproj/io/html/LandingHTMLModule.kt
@@ -1,10 +1,10 @@
-package ru.hse.sd.hwproj.server.html
+package ru.hse.sd.hwproj.io.html
 
 import io.ktor.server.application.*
 import io.ktor.server.html.*
 import io.ktor.server.routing.*
-import ru.hse.sd.hwproj.server.html.student.makeStudentLandingPage
-import ru.hse.sd.hwproj.server.html.teacher.makeTeacherLandingPage
+import ru.hse.sd.hwproj.io.html.student.makeStudentLandingPage
+import ru.hse.sd.hwproj.io.html.teacher.makeTeacherLandingPage
 
 fun Application.addLandingHTMLModule() {
     routing {
