+++ src/main/kotlin/ru/hse/sd/hwproj/io/html/student/StudentHTMLModule.kt
@@ -1,4 +1,4 @@
-package ru.hse.sd.hwproj.server.html.student
+package ru.hse.sd.hwproj.io.html.student
 
 import io.ktor.server.application.*
 import io.ktor.server.html.*
@@ -6,8 +6,8 @@ import io.ktor.server.request.*
 import io.ktor.server.routing.*
 import ru.hse.sd.hwproj.interactor.Interactor
 import ru.hse.sd.hwproj.models.*
-import ru.hse.sd.hwproj.server.html.makeListSubmissionsPage
-import ru.hse.sd.hwproj.server.html.makeSubmissionDetailsPage
+import ru.hse.sd.hwproj.io.html.makeListSubmissionsPage
+import ru.hse.sd.hwproj.io.html.makeSubmissionDetailsPage
 
 fun Application.addStudentHTMLModule(interactor: Interactor) {
     routing {
