+++ src/main/kotlin/ru/hse/sd/hwproj/models/RequestModels.kt
@@ -1,4 +1,5 @@
 @file:UseSerializers(TimestampSerializer::class)
+@file:Suppress("CanSealedSubClassBeObject")
 
 package ru.hse.sd.hwproj.models
